Added empty component class, changed entity behavior
Entity now uses default empty components rather than if-checking.
